DiscoveryWorker: move drain signal handling out
